+++ weevent-processor/src/main/java/com/webank/weevent/processor/utils/CommonUtil.java
@@ -412,152 +412,6 @@ public class CommonUtil {
         return rules.getKey().getFromDestination().equals(rules.getValue().getFromDestination());
     }
 
-
-    /**
-     * count the position, for version 1.2 abs and ceil and floor and round
-     *
-     * @param conditionField original condition message
-     * @param sb condition buffer
-     * @return amount
-     */
-    private static int changePosition(String conditionField, String sb) {
-        int changePosition = 0;
-        if (conditionField.length() > sb.length()) {
-            changePosition = conditionField.length() - sb.length();
-        }
-        return changePosition;
-    }
-
-    public static String[][] stringConvertArray(String s) {
-        String[] s1 = s.replaceAll("],", "]#").split("#");
-        String[][] arr = new String[s1.length][];
-        for (int i = 0; i < arr.length; i++) {
-            String[] s2 = s1[i].split(",");
-            arr[i] = new String[s2.length];
-            for (int j = 0; j < s2.length; j++) {
-                arr[i][j] = s2[j].replaceAll("\\[|\\]", "").replace("\"", "").trim();
-            }
-        }
-        return arr;
-    }
-
-    public static String analysisSystemFunction(String[][] systemFunctionMessage, String payload, String
-            conditionField) throws IOException {
-
-        Map maps = JsonUtil.parseObjectToMap(payload);
-        Map<String, Object> payloadMap = new ConcurrentHashMap<>();
-        for (Object map : maps.entrySet()) {
-            payloadMap.put((String) ((Map.Entry) map).getKey(), ((Map.Entry) map).getValue());
-        }
-        return replaceCondition(systemFunctionMessage, conditionField, payloadMap);
-    }
-
-    public static Pair<StringBuilder, Integer> replaceCase(StringBuilder sb, String conditionField, Map payload, String[] arr, int changePosition) {
-        String type = arr[2];
-        String left = arr[3]; // end position
-        String middle = "";
-        String right = "";
-
-        if (arr.length == 5) {
-            left = arr[3];
-            right = arr[4];
-        } else if (arr.length == 6) { // multi parameter
-            left = arr[3];
-            middle = arr[4];
-            right = arr[5];
-        }
-
-        String replaceContent = "";
-        Integer start = Integer.valueOf(arr[0]);
-        Integer end = Integer.valueOf(arr[1]);
-        switch (type) {
-//            case "now":
-//                sb.replace(start - changePosition, end - changePosition, String.valueOf(new Date().getTime()));
-//                break;
-//            case "currentDate":
-//                String date = new SimpleDateFormat("yyyyMMdd").format(new Date());
-//                sb.replace(start - changePosition, end - changePosition, date);
-//                break;
-//            case "currentTime":
-//                String time = new SimpleDateFormat("HHmmss").format(new Date());
-//                sb.replace(start - changePosition, end - changePosition, time);
-//                break;
-
-            case "abs":
-                sb.replace(start - changePosition, end - changePosition, String.valueOf(Math.abs((Integer) payload.get(arr[3]))));
-                changePosition = changePosition(conditionField, sb.toString());
-                break;
-
-            case "ceil":
-                sb.replace(start - changePosition, end - changePosition, String.valueOf(Math.ceil((Integer) payload.get(arr[3]))));
-                changePosition = changePosition(conditionField, sb.toString());
-                break;
-
-            case "floor":
-                sb.replace(start - changePosition, end - changePosition, String.valueOf(Math.floor((Integer) payload.get(arr[3]))));
-                changePosition = changePosition(conditionField, sb.toString());
-                break;
-
-            case "round":
-                sb.replace(start - changePosition, end - changePosition, String.valueOf(Math.round(Math.round((Integer) payload.get(arr[3])))));
-                changePosition = changePosition(conditionField, sb.toString());
-                break;
-
-            case "substring":
-                if (!"".equals(middle)) {
-                    replaceContent = "\"" + payload.get(left).toString().substring(Integer.valueOf(middle), Integer.valueOf(right)) + "\"";
-                    sb.replace(start - changePosition, end - changePosition, replaceContent);
-                } else {
-                    replaceContent = "\"" + payload.get(left).toString().substring(Integer.valueOf(right)) + "\"";
-                    sb.replace(start - changePosition, end - changePosition, replaceContent);
-                }
-                changePosition = changePosition(conditionField, sb.toString());
-                break;
-
-            case "concat":
-                replaceContent = "\"" + payload.get(left).toString().concat(payload.get(right).toString()) + "\"";
-                sb.replace(start - changePosition, end - changePosition, replaceContent);
-                changePosition = changePosition(conditionField, sb.toString());
-                break;
-
-            case "trim":
-                replaceContent = "\"" + payload.get(left).toString().trim() + "\"";
-                sb.replace(start - changePosition, end - changePosition, replaceContent);
-                changePosition = changePosition(conditionField, sb.toString());
-                break;
-
-            case "lcase":
-                replaceContent = "\"" + payload.get(left).toString().toLowerCase() + "\"";
-                sb.replace(start - changePosition, end - changePosition, replaceContent);
-                changePosition = changePosition(conditionField, sb.toString());
-                break;
-
-            default:
-                log.info("conditionField:{}", conditionField);
-                break;
-        }
-        return new Pair<>(sb, changePosition);
-    }
-
-    /**
-     * @param arr enhance function message
-     * @param conditionField original condition details
-     * @param payload payload
-     * @return condition
-     */
-    public static String replaceCondition(String[][] arr, String conditionField, Map payload) {
-        StringBuilder sb = new StringBuilder(conditionField);
-        int changePosition = 0;
-
-        for (int i = 0; i < arr.length; i++) {
-            Pair<StringBuilder, Integer> ret = replaceCase(sb, conditionField, payload, arr[i], changePosition);
-            sb = ret.getKey();
-            changePosition = ret.getValue();
-        }
-        log.info("sb:{}", sb);
-        return sb.toString();
-    }
-
     public static boolean isDate(String strDate) {
         Pattern pattern = Pattern
                 .compile("^((\\d{2}(([02468][048])|([13579][26]))[\\-\\/\\s]?((((0?[13578])|(1[02]))[\\-\\/\\s]?((0?[1-9])|([1-2][0-9])|(3[01])))|(((0?[469])|(11))[\\-\\/\\s]?((0?[1-9])|([1-2][0-9])|(30)))|(0?2[\\-\\/\\s]?((0?[1-9])|([1-2][0-9])))))|(\\d{2}(([02468][1235679])|([13579][01345789]))[\\-\\/\\s]?((((0?[13578])|(1[02]))[\\-\\/\\s]?((0?[1-9])|([1-2][0-9])|(3[01])))|(((0?[469])|(11))[\\-\\/\\s]?((0?[1-9])|([1-2][0-9])|(30)))|(0?2[\\-\\/\\s]?((0?[1-9])|(1[0-9])|(2[0-8]))))))(\\s(((0?[0-9])|([1-2][0-3]))\\:([0-5]?[0-9])((\\s)|(\\:([0-5]?[0-9])))))?$");
